+++ example/dcgan.go
@@ -1,100 +1,23 @@
 package main
 
 import (
-	"compress/gzip"
 	"fmt"
-	"io"
 	"log"
-	"net/http"
-	"os"
-	"os/user"
-	"path"
 
 	torch "github.com/wangkuiyi/gotorch"
 	nn "github.com/wangkuiyi/gotorch/nn"
 	F "github.com/wangkuiyi/gotorch/nn/functional"
 	"github.com/wangkuiyi/gotorch/nn/initializer"
+	"github.com/wangkuiyi/gotorch/vision"
 )
 
-const (
-	images = "train-images-idx3-ubyte"
-	labels = "train-labels-idx1-ubyte"
-	site   = "http://yann.lecun.com/exdb/mnist/"
-)
-
-func downloadMNIST() error {
-	if e := downloadIfNotYet(images); e != nil {
-		return e
-	}
-	if e := downloadIfNotYet(labels); e != nil {
-		return e
-	}
-	return nil
-}
-
-func downloadIfNotYet(fn string) error {
-	f := path.Join(dataDir(), fn)
-	if !fileExists(f) {
-		if e := download(site+fn+".gz", f); e != nil {
-			return e
-		}
-	}
-	return nil
-}
-
-func dataDir() string {
-	u, e := user.Current()
-	if e != nil {
-		return "testdata/mnist"
-	}
-	return path.Join(u.HomeDir, ".cache/mnist")
-}
-
-func fileExists(fn string) bool {
-	info, e := os.Stat(fn)
-	if os.IsNotExist(e) {
-		return false
-	}
-	return !info.IsDir()
-}
-
-func download(url, fn string) error {
-	resp, e := http.Get(url)
-	if e != nil {
-		return e
-	}
-	defer resp.Body.Close()
-
-	r, e := gzip.NewReader(resp.Body)
-	if e != nil {
-		return e
-	}
-	defer r.Close()
-
-	if e := os.MkdirAll(dataDir(), 0744); e != nil {
-		return e
-	}
-
-	f, e := os.Create(fn)
-	if e != nil {
-		return e
-	}
-	defer f.Close()
-
-	_, e = io.Copy(f, r)
-	if e != nil && e != io.EOF {
-		return e
-	}
-
-	return nil
-}
-
+// Generator network
 type Generator struct {
 	nn.Module
 	Main *nn.SequentialModule
 }
 
-func NewGenerator(nz int64) *Generator {
+func newGenerator(nz int64) *Generator {
 	g := &Generator{
 		Main: nn.Sequential(
 			nn.ConvTranspose2d(nz, 256, 4, 1, 0, 0, 1, false, 1, "zero"),
@@ -114,16 +37,18 @@ func NewGenerator(nz int64) *Generator {
 	return g
 }
 
+// Forward method
 func (g *Generator) Forward(x torch.Tensor) torch.Tensor {
 	return g.Main.Forward(x).(torch.Tensor)
 }
 
+// Discriminator network
 type Discriminator struct {
 	nn.Module
 	Main *nn.SequentialModule
 }
 
-func NewDiscriminator() *Discriminator {
+func newDiscriminator() *Discriminator {
 	d := &Discriminator{
 		Main: nn.Sequential(
 			nn.Conv2d(1, 64, 4, 2, 1, 1, 1, false, "zeros"),
@@ -142,22 +67,23 @@ func NewDiscriminator() *Discriminator {
 	return d
 }
 
+// Forward method
 func (d *Discriminator) Forward(x torch.Tensor) torch.Tensor {
 	return d.Main.Forward(x).(torch.Tensor).View([]int64{-1, 1}).Squeeze(1)
 }
 
 func main() {
-	if e := downloadMNIST(); e != nil {
+	if e := vision.DownloadMNIST(); e != nil {
 		log.Printf("Cannot find or download MNIST dataset: %v", e)
 	}
 	transforms := []torch.Transform{torch.NewNormalize(0.5, 0.5)}
-	mnist := torch.NewMNIST(dataDir(), transforms)
+	mnist := torch.NewMNIST(vision.MNISTDir(), transforms)
 
 	nz := int64(100)
 	lr := 0.0002
 
-	netG := NewGenerator(nz)
-	netD := NewDiscriminator()
+	netG := newGenerator(nz)
+	netD := newDiscriminator()
 
 	optimizerD := torch.Adam(lr, 0.5, 0.5, 0.0)
 	optimizerD.AddParameters(netD.Parameters())
@@ -183,7 +109,6 @@ func main() {
 			output := netD.Forward(batch.Data)
 			errDReal := F.BinaryCrossEntropy(output, label, torch.Tensor{}, "mean")
 			errDReal.Backward()
-			DX := output.Mean().Item()
 
 			// train with fake
 			noise := torch.RandN([]int64{batch.Data.Shape()[0], nz, 1, 1}, false)
@@ -192,7 +117,6 @@ func main() {
 			output = netD.Forward(fake.Detach())
 			errDFake := F.BinaryCrossEntropy(output, label, torch.Tensor{}, "mean")
 			errDFake.Backward()
-			DGZ1 := output.Mean().Item()
 			errD := errDReal.Item() + errDFake.Item()
 			optimizerD.Step()
 
@@ -202,10 +126,9 @@ func main() {
 			output = netD.Forward(fake)
 			errG := F.BinaryCrossEntropy(output, label, torch.Tensor{}, "mean")
 			errG.Backward()
-			DGZ2 := output.Mean().Item()
 			optimizerG.Step()
-			fmt.Printf("[%d/%d][%d] Loss_D: %f Loss_G: %f D(x): %f D(G(z)): %f / %f\n",
-				epoch, epochs, i, errD, errG.Item(), DX, DGZ1, DGZ2)
+			fmt.Printf("[%d/%d][%d] Loss_D: %f Loss_G: %f\n",
+				epoch, epochs, i, errD, errG.Item())
 			if i%checkpointStep == 0 {
 				samples := netG.Forward(torch.RandN([]int64{10, nz, 1, 1}, false))
 				ckName := fmt.Sprintf("dcgan-sample-%d.pt", checkpointCount)
